Extract library-type inference to importer_library_types
Sprint 4, PR #4.  Second peel of modules/importer.py.  Moves the
~283-line cluster that answers 'given a user's uploaded config.yml,
what kind of library is each Plex library and which
collections/overlays can it support?'.

Public API in modules/importer_library_types:
- infer_library_types(config_data)
    Given a config.yml dict, return (library_types, inference_list).
    library_types maps library names to types (movie/show/music/...);
    inference_list describes HOW each guess was made (explicit
    'library_type' key, from operations/collections present, or as
    a fallback).
- build_library_type_plan(config_data, movie_names, show_names)
    Combines infer_library_types with the caller-provided
    Plex-detected library-name sets to produce the final
    (library_types, library_inference, needs_confirmation)
    triple used by the import-preview UI.
- normalize_library_type(value)
    Public wrapper around _normalize_library_type -- canonicalizes
    a raw library_type string from user YAML.

Internal '_' helpers moved with them:
- _normalize_library_type   -- canonicalization
- _build_collection_index   -- load + index quickstart_collections.json
- _build_overlay_index      -- load + index quickstart_overlays.json
- _resolve_collection_id    -- match user default against index
- _resolve_overlay_id       -- match user default against index

This is a PURE MOVE.  No inference logic changed, no branches
reordered, no re-normalizations added.  Every returned
(library_types, inference_list, needs_confirmation) triple is
byte-identical.

Compatibility:
- All 8 names re-exported from modules/importer.
  * build_library_type_plan is called externally by
    blueprints/import_config_routes.py L639
  * _build_collection_index and _build_overlay_index are still
    called by the mega prepare_import_payload (which stayed
    behind in importer.py) at L730-731 -- they resolve via the
    re-export.
  * The other five are kept accessible via importer.X for
    future test monkeypatching / debugging.

- _build_attribute_sets, which sat between the index builders
  and the library-type functions in the source file, is a
  SEPARATE concern (parses attribute config metadata) and stays
  in importer.py with its sole caller prepare_import_payload.

Dependency direction: importer -> importer_library_types
(one way).  The new module has zero imports from importer, so
no circular imports possible.

Impact:
- modules/importer.py: 1777 -> 1494 (-283 / -15.9%)
- modules/importer_library_types.py: NEW ~340 lines

All 1285 tests pass.  Ruff + black clean.  Smoke-tested
normalize_library_type + infer_library_types + build_library_type_plan
end-to-end -- pipeline correctly returns {'Movies': 'movie',
'TV Shows': 'show'} for the standard case.
